# swagger.yaml
openapi: 3.0.0
info:
    title: Management User System
    description: This is a RESTful API of Management User System. When you login to system, you got a token. And insert token to Authorize button.
    termsOfService: "https://github.com/offer-14m"
    contact:
        email: "off14mteam@swagger.io"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    version: 3.0.0
servers:
    - url: http://localhost:1996
      description: Development server
...
paths:
...
    /api/users/register:
        post:
            tags:
                - users
            summary: Register a new user
            requestBody:
                description: Register a new user linked to an employee to get the sufficient permissions to use the app
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserRegister"
            responses:
                200:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    message:
                                        type: string
                                        example: Login successfully
                401:
                    description: Unauthorization
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Unauthorization"

                500:
                    description: server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: Failed to register.
    /api/users/login:
        post:
            tags:
                - users
            summary: Login user
            security:
                - bearerAuth: []
            requestBody:
                description: User login to use app
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserLogin"
            responses:
                200:
                    description: Login successfull
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    message:
                                        type: string
                                        example: Login successfully
                                    data:
                                        type: object
                                        properties:
                                            success:
                                                type: boolean
                                                example: true
                                            token:
                                                type: string
                                                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbXBsb3llZU51bWJlciI6MTAwMiwib2ZmaWNlQ29kZSI6IjEiLCJqb2JUaXRsZSI6IlByZXNpZGVudCIsImlhdCI6MTYzNTk2NjQ5MCwiZXhwIjoxNjM1OTcwMDkwfQ.0laYWT2XaNYJz2q6ISvcjQNO8risVnF4VCbhr5hWqiQ
                404:
                    description: Login failed
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: string
                                        example: false
                                    message:
                                        type: string
                                        example: Invalid username or password.
                500:
                    description: server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: Failed to Login.
...
components:
    schemas:
        UserRegister:
            type: object
            properties:
                username:
                    type: string
                    minLength: 3
                    maxLength: 20
                password:
                    type: string
                    minLength: 6
                    maxLength: 100
                    pattern: "^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{6,}$"
                employeeNumber:
                    type: number
                    minimum: 1
            required:
                - username
                - password
                - employeeNumber
        UserLogin:
            type: object
            properties:
                username:
                    type: string
                    minLength: 3
                    maxLength: 20
                    example: hoamx1
                password:
                    type: string
                    minLength: 6
                    maxLength: 100
                    pattern: "^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{6,}$"
                    example: hoa12345@President
            required:
                - username
                - password
...
        Unauthorization:
            type: object
            properties:
                err:
                    type: string
                    example: Unauthorization
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT